Fix uninitialized bug in recursive_delete_files and turn of some warnings

+++ src/scaffolding/bin_hash.c
@@ -8,7 +8,7 @@ float get_avg_barcode(struct asm_graph_t *g)
 	int count = 0;
 	uint64_t sum = 0;
 	for (int i = 0; i < g->n_e; ++i) {
-		if (get_edge_len(&g->edges[i]) > global_count_bc_size) {
+		if (get_edge_len(&g->edges[i]) > (uint32_t) global_count_bc_size) {
 			sum += g->edges[i].barcodes_scaf.n_item;
 			count++;
 		}
